requirements.txt
django
openai
pymysql
cryptography
requests
djangorestframework
djangorestframework-simplejwt
pymdf